feat: 完成标准合规性修复和代码结构优化
Phase 1 (紧急修复):
- 统一数据库配置: 删除重复的db.py，统一使用database.py
- 统一依赖注入: 将所有get_session()改为get_db()，确保一致性
- 添加API版本号: 为所有API路径添加/v1/版本号前缀，符合RESTful标准
- 路由前缀统一: 统一所有路由前缀为/api/v1，解决多种前缀模式并存问题

Phase 2 (标准化修复):
- 文件重命名: routes.py -> api_routes.py, models.py -> data_models.py
- 主键类型修复: 修改所有主键类型从Integer改为BigInteger，支持大数据量
- 代码注释规范: 补充code-standards.md中的文件头部注释强制规范

修复问题:
- 解决数据库配置重复和依赖注入不一致问题
- 修复API版本缺失和路由前缀混乱问题
- 统一文件命名规范，提升代码可维护性
- 升级数据库主键类型，为大数据量做准备

本次提交通过新建立的强制检查机制完成，确保所有修改都符合标准文档要求。

+++ app/data_models.py
@@ -1,5 +1,5 @@
 from sqlalchemy.orm import declarative_base
-from sqlalchemy import Column, Integer, String, Text, DECIMAL, DateTime, Boolean, ForeignKey, Index
+from sqlalchemy import Column, Integer, BigInteger, String, Text, DECIMAL, DateTime, Boolean, ForeignKey, Index
 from sqlalchemy.sql import func
 from sqlalchemy.orm import relationship
 
@@ -8,7 +8,7 @@ Base = declarative_base()
 class User(Base):
     __tablename__ = 'users'
     
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     username = Column(String(50), unique=True, nullable=False)
     email = Column(String(200), unique=True, nullable=False)
     # 密码认证
@@ -35,7 +35,7 @@ class User(Base):
 class Category(Base):
     __tablename__ = 'categories'
     
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     name = Column(String(100), nullable=False)
     parent_id = Column(Integer, ForeignKey('categories.id'), nullable=True)
     sort_order = Column(Integer, default=0)
@@ -49,7 +49,7 @@ class Category(Base):
 class Product(Base):
     __tablename__ = 'products'
     
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     name = Column(String(200), nullable=False)
     sku = Column(String(100), unique=True, nullable=False)
     description = Column(Text, nullable=True)
@@ -90,7 +90,7 @@ class Product(Base):
 class Order(Base):
     __tablename__ = 'orders'
     
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     order_no = Column(String(32), unique=True, nullable=False)  # 订单号
     
     # 用户关联
@@ -134,7 +134,7 @@ class Order(Base):
 class OrderItem(Base):
     __tablename__ = 'order_items'
     
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     
     # 关联
     order_id = Column(Integer, ForeignKey('orders.id'), nullable=False)
@@ -165,7 +165,7 @@ class OrderItem(Base):
 # 保留现有的 Certificate 模型（暂时保留，后续可能重构为质量认证系统的一部分）
 class Certificate(Base):
     __tablename__ = 'certificates'
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     name = Column(String(200), nullable=False)
     issuer = Column(String(200), nullable=True)
     serial = Column(String(200), unique=True, nullable=False)
@@ -175,7 +175,7 @@ class Payment(Base):
     """支付单模型 - V1.0 Mini-MVP"""
     __tablename__ = 'payments'
     
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     order_id = Column(Integer, ForeignKey('orders.id'), nullable=False)
     user_id = Column(Integer, ForeignKey('users.id'), nullable=False)  # 所有权字段
     
@@ -230,7 +230,7 @@ class Refund(Base):
     """退款单模型 - V1.0 Mini-MVP"""
     __tablename__ = 'refunds'
     
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     payment_id = Column(Integer, ForeignKey('payments.id'), nullable=False)
     
     # 退款信息
@@ -263,7 +263,7 @@ class Cart(Base):
     """购物车模型"""
     __tablename__ = 'carts'
     
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     user_id = Column(Integer, ForeignKey('users.id'), nullable=False)
     status = Column(String(20), default='active', nullable=False)  # active, inactive, converted
     total_amount = Column(DECIMAL(10, 2), default=0.00)
@@ -288,7 +288,7 @@ class CartItem(Base):
     """购物车项目模型"""
     __tablename__ = 'cart_items'
     
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     cart_id = Column(Integer, ForeignKey('carts.id'), nullable=False)
     product_id = Column(Integer, ForeignKey('products.id'), nullable=False)
     quantity = Column(Integer, nullable=False)
@@ -341,7 +341,7 @@ class Inventory(Base):
     """库存主表"""
     __tablename__ = "inventory"
 
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     product_id = Column(
         Integer, 
         ForeignKey("products.id", ondelete="CASCADE"), 
@@ -389,7 +389,7 @@ class InventoryTransaction(Base):
     """库存变动记录表"""
     __tablename__ = "inventory_transactions"
 
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     product_id = Column(Integer, ForeignKey("products.id", ondelete="CASCADE"), nullable=False)
     transaction_type = Column(ENUM(TransactionType), nullable=False)
     quantity = Column(Integer, nullable=False)
@@ -417,7 +417,7 @@ class CartReservation(Base):
     """购物车库存预占表"""
     __tablename__ = "cart_reservations"
 
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     user_id = Column(Integer, ForeignKey("users.id", ondelete="CASCADE"), nullable=False)
     product_id = Column(Integer, ForeignKey("products.id", ondelete="CASCADE"), nullable=False)
     reserved_quantity = Column(Integer, nullable=False)
